Added new production encryption keys
Added new encrypted property keys to be used by the CI environment
and by future cloud service and the new production master key
to be able to decrypt the secret property keys for security database,
openbank database, ssl client/server that CI/cloud environment is only
aware of.

Updated Testcontainers to version 1.15.2.

Refactored the SSL encrypted string to be able to execute integration
tests locally and for the test production environment.